Rename redux folder into data folder
- data json into data folder
- sort func into utils
- update imports
