+++ src/components/FamiliarCalc/index.jsx
@@ -16,7 +16,7 @@ import TableContainer from '@mui/material/TableContainer';
 import TableHead from '@mui/material/TableHead';
 import TableRow from '@mui/material/TableRow';
 import Paper from '@mui/material/Paper';
-import familiarsData from '../../redux/familiars.json';
+import familiarsData from '../../data/familiars.json';
 
 import styles from "./index.module.css";
 import { VerticalSpacing } from "../../utils/utils";
