+++ src/data/calcOptions.json
