+++ src/utils/utils.js
@@ -1,40 +1,44 @@
-
 export const VerticalSpacing = 3;
 
 export function calcIF(vals) {
-    const total = 100 +
-        vals.rune1 +
-        vals.rune2 +
-        vals.guild +
-        vals.consumable +
-        vals.adgor +
-        vals.daily;
-    return (total * vals.encounter).toFixed(2);
+  const total =
+    100 +
+    vals.rune1 +
+    vals.rune2 +
+    vals.guild +
+    vals.consumable +
+    vals.adgor +
+    vals.daily;
+  return (total * vals.encounter).toFixed(2);
 }
 
 export function getIFEquation(vals) {
-    return `(base:100 + rune:${vals.rune1} + rune:${vals.rune2} + guild:${vals.guild} + consumable:${vals.consumable} + adgor:${vals.adgor} + daily:${vals.daily}) * (1 + encounter:${(vals.encounter-1).toFixed(2)})`;
+  return `(base:100 + rune:${vals.rune1} + rune:${vals.rune2} + guild:${vals.guild} + consumable:${vals.consumable} + adgor:${vals.adgor} + daily:${vals.daily}) * (1 + encounter:${(vals.encounter - 1).toFixed(2)})`;
 }
 
 export function cleanVal(val) {
-    if (typeof val === "number") {
-        return val;
-    }
-    return Number.parseFloat(val);
+  if (typeof val === "number") {
+    return val;
+  }
+  return Number.parseFloat(val);
 }
 
 export function getOptionValue(section, key) {
-    if (section.options) {
-        return section.options[key]?.value ?? 0;
+  if (section.options) {
+    return section.options[key]?.value ?? 0;
+  }
+  for (const group of Object.values(section.groups)) {
+    if (key in group) {
+      return group[key].value;
     }
-    for (const group of Object.values(section.groups)) {
-        if (key in group) {
-            return group[key].value;
-        }
-    }
-    return 0;
+  }
+  return 0;
 }
 
 export function getEncounterIFForDisplay(val) {
-    return (val-1).toFixed(2)*100;
+  return (val - 1).toFixed(2) * 100;
+}
+
+export function naturalSort(a, b) {
+  return a.localeCompare(b, undefined, { numeric: true, sensitivity: "base" });
 }
